ui_prevision/requirements.txt
streamlit
pandas
numpy
python-dateutil
scikit-learn
joblib
glob2
requests
xgboost
mlflow 
prefect